Add projection subqueries to top level cypher resolver
