+++ packages/graphql/src/translate/index.ts
@@ -22,3 +22,4 @@ export { translateRead } from "./translate-read";
 export { default as translateUpdate } from "./translate-update";
 export { translateDelete } from "./translate-delete";
 export { default as translateAggregate } from "./translate-aggregate";
+export { translateTopLevelCypher } from "./translate-top-level-cypher";
